src/test/javascript/spec/app/entities/regroupement-regional/regroupement-regional.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { RegroupementRegionalService } from 'app/entities/regroupement-regional/regroupement-regional.service';
import { IRegroupementRegional, RegroupementRegional } from 'app/shared/model/regroupement-regional.model';

describe('Service Tests', () => {
  describe('RegroupementRegional Service', () => {
    let injector: TestBed;
    let service: RegroupementRegionalService;
    let httpMock: HttpTestingController;
    let elemDefault: IRegroupementRegional;
    let expectedResult: IRegroupementRegional | IRegroupementRegional[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(RegroupementRegionalService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new RegroupementRegional(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a RegroupementRegional', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new RegroupementRegional()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a RegroupementRegional', () => {
        const returnedFromService = Object.assign(
          {
            sigle: 'BBBBBB',
            nomRegroupement: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of RegroupementRegional', () => {
        const returnedFromService = Object.assign(
          {
            sigle: 'BBBBBB',
            nomRegroupement: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a RegroupementRegional', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
